Move NotNull rescue to Validations
